+++ R/samc.R
@@ -131,10 +131,6 @@ setMethod(
   function(data, absorption, fidelity, tr_args) {
     .validate_tr_args(tr_args)
 
-    data = .rasterize(data)
-    absorption = .rasterize(absorption)
-    fidelity = .rasterize(fidelity)
-
     tr_fun <- tr_args$fun
     directions <-tr_args$dir
     symm <- tr_args$sym
@@ -233,9 +229,9 @@ setMethod(
             tr_args = "list"),
   function(data, absorption, fidelity, tr_args) {
 
-    data = .rasterize(data)
-    absorption = .rasterize(absorption)
-    fidelity = .rasterize(fidelity)
+    data = rasterize(data)
+    absorption = rasterize(absorption)
+    fidelity = rasterize(fidelity)
 
     samc_obj = samc(data, absorption, fidelity, tr_args = tr_args)
     samc_obj@source = "RasterLayer"
@@ -267,8 +263,8 @@ setMethod(
             tr_args = "list"),
   function(data, absorption, tr_args) {
 
-    data = .rasterize(data)
-    absorption = .rasterize(absorption)
+    data = rasterize(data)
+    absorption = rasterize(absorption)
 
     samc_obj = samc(data, absorption, tr_args = tr_args)
     samc_obj@source = "RasterLayer"
@@ -285,9 +281,9 @@ setMethod(
             tr_args = "list"),
   function(data, absorption, fidelity, tr_args) {
 
-    data <- .rasterize(data)
-    absorption <- .rasterize(absorption)
-    fidelity <- .rasterize(fidelity)
+    data <- rasterize(data)
+    absorption <- rasterize(absorption)
+    fidelity <- rasterize(fidelity)
 
     samc_obj = samc(data, absorption, fidelity, tr_args)
     samc_obj@source = "matrix"
@@ -306,8 +302,8 @@ setMethod(
             tr_args = "list"),
   function(data, absorption, tr_args) {
 
-    data <- .rasterize(data)
-    absorption <- .rasterize(absorption)
+    data <- rasterize(data)
+    absorption <- rasterize(absorption)
 
     samc_obj = samc(data, absorption, tr_args = tr_args)
     samc_obj@source = "matrix"